improve error handling for config options
When the option fields for a config option could not be parsed, the
build failed with an error message about the "shortdesc" option, even
though the problem might be in any of the fields.
Fix the error handling to give a clearer message, and to not break
the build but just leave out the problematic option.

Signed-off-by: Ruth Fuchss <ruth.fuchss@canonical.com>