add a custom `none` role that is excluded from the output
To ignore woke exceptions, you must provide a comment (see
https://docs.getwoke.tech/ignore/#in-line-and-next-line-ignoring ).
However, RST doesn't really have a concept of comments, and the
workaround that is commonly used for comments requires empty lines,
which again doesn't work for wokeignore.

The `:none:` role makes it possible to specify the wokeignore
comment in the input file, but exclude it from the output.

Signed-off-by: Ruth Fuchss <ruth.fuchss@canonical.com>