config-options: fix JavaScript selector
The "expand all" button wasn't always displayed correctly. This is
because the "first-of-type" selector isn't meant to distinguish
between classes (only types).
So instead of using "first-of-type", select all elements with the
".configoption" class and then undo the addition of the "expand all"
button for all elements that are preceded by another ".configoption"
element.

Fixes #35

Signed-off-by: Ruth Fuchss <ruth.fuchss@canonical.com>