Remove :state: option
It doesn't make any sense to have this as any unsupported release has no
corresponding directories on cdimage and so we can never produce output
for them anyway.

We also have the :empty: option now to include messages in the case that
no supported releases are found (e.g. if the devel release is the only
matching one).

Fixes: #53