Refine the interpretation of the supported flag
The "supported" flag in changelogs.ubuntu.com/meta-release doesn't
actually mean "supported", but rather "will be offered for upgrade". For
instance, immediately after an LTS release, this is 0 until the first
point-release, after which it is moved to 1. It may also be set to 0 for
other reasons (like in plucky where an upgrade issue resulted in it
being initially disabled).

However, in both cases we should still be listing the images. This
commit refines the interpretation of "supported" to be "within the
life-span of the release (according to release date and LTS state), or
explicitly marked upgradable".

Also took the opportunity to make all datetime's tz-aware (naive
datetime's are due to be deprecated in future Python)

Fixes: #57